# perf: remove `camelCaseProperty` call in `renderRule()` inside Fela