#### Pull request checklist

- [ ] Addresses an existing issue: Fixes #0000
- [ ] Include a change request file using `$ yarn change`

#### Description of changes

[`renderRule()`](http://fela.js.org/docs/basics/Renderer.html#renderrule) is a hot function for us as we are calling it on each our cache miss. A callstack for Fela looks like this:

```
renderRule(rule, props)
  _renderStyle((rule, props)) // styles are passed resolved; performs plugins calls
    _renderStyleToClassNames(style) // performs Fela cache lookups
````

Inside `_renderStyleToClassNames()` exists a call to `generateDeclarationReference()` what creates a lookup key for Fela cache:

https://github.com/robinweser/fela/blob/97148419e7941dbd5867b35df6ac4de6c5ed3e55/packages/fela/src/createRenderer.js#L269-L277

```js
          const declarationReference = generateDeclarationReference(
            property,
            value,
            pseudo,
            media,
            support
          )

          if (!renderer.cache.hasOwnProperty(declarationReference)) {
```

Inside `generateDeclarationReference()` there is call to [`camelCaseProperty()`](https://github.com/robinweser/css-in-js-utils/blob/master/modules/camelCaseProperty.js) which performs a transform to force camelcase via regexps (and actually this makes this function slow). This is visible in our perfs and in my test environment:

```
- Fela                                        69346 ops/sec
- Fela without `camelCaseProperty()`         261801 ops/sec
```

<details>
 <summary>Test rule</summary>

```js
const fullRule = () => ({
  backgroundColor: "transparent",
  backgroundPosition: "center",
  display: "flex",
  float: 0,
  left: 0,
  right: 0,
  flexGrow: 1,
  marginLeft: 0,
  marginRight: 0,
  marginBottom: 0,
  marginTop: 0,
  paddingLeft: 0,
  paddingRight: 0,
  paddingBottom: 0,
  paddingTop: 0,
  outline: 0,
  opacity: 1,
  height: "20px",
  maxHeight: "20px",
  minWidth: "20px",
  width: "20px",
  visibility: "hidden",
  zIndex: 1,
});
```

</details>

------

__Sad part:__ This will break rehydration process from SSR content, but as we are not using it we are fine to go.

------

## Benchmarks 🚀 

A quick improvement summary, all data from benchmarks are available below:

### `new-page`
| Example                      | Improvement |
| ---------------------------- | ----------- |
| ChatWithPopover.perf.tsx     | -8%         |
| DropdownManyItems.perf.tsx   | -12%        |
| TreeWith60ListItems.perf.tsx | -4%         |
### `same-page`
| Example                      | Improvement |
| ---------------------------- | ----------- |
| ChatWithPopover.perf.tsx     | -15%        |
| DropdownManyItems.perf.tsx   | -23%        |
| TreeWith60ListItems.perf.tsx | -24%        |

<details>
  <summary>All data from benchmarks</summary>

# `master`

### `new-page`
| Example                      | min    | avg    | median | max    |
| ---------------------------- | ------ | ------ | ------ | ------ |
| ChatWithPopover.perf.tsx     | 237.22 | 248.88 | 246.91 | 308.35 |
| ChatWithPopover.perf.tsx     | 229.51 | 246.61 | 243.15 | 325.39 |
| DropdownManyItems.perf.tsx   | 252.27 | 263.55 | 261.07 | 323.86 |
| DropdownManyItems.perf.tsx   | 245.89 | 257.4  | 254.37 | 312.04 |
| TreeWith60ListItems.perf.tsx | 146.86 | 156.25 | 155.14 | 187.97 |
| TreeWith60ListItems.perf.tsx | 135.93 | 141.01 | 140    | 167.06 |
### `same-page`
| Example                      | min    | avg    | median | max    |
| ---------------------------- | ------ | ------ | ------ | ------ |
| ChatWithPopover.perf.tsx     | 109.36 | 128.45 | 122.3  | 248.05 |
| ChatWithPopover.perf.tsx     | 108.77 | 125.84 | 118.78 | 253.32 |
| DropdownManyItems.perf.tsx   | 120.85 | 137.22 | 131.49 | 234.47 |
| DropdownManyItems.perf.tsx   | 123.85 | 138.49 | 132.56 | 236.83 |
| TreeWith60ListItems.perf.tsx | 57.13  | 74.35  | 70.11  | 138.08 |
| TreeWith60ListItems.perf.tsx | 58.71  | 72.57  | 67.63  | 151.7  |

---

# With changes

### `new-page`
| Example                      | min    | avg    | median | max    |
| ---------------------------- | ------ | ------ | ------ | ------ |
| ChatWithPopover.perf.tsx     | 218.36 | 228.24 | 225.52 | 271.82 |
| ChatWithPopover.perf.tsx     | 217.68 | 229.73 | 227.19 | 280.93 |
| DropdownManyItems.perf.tsx   | 216.5  | 231.2  | 227.94 | 268.09 |
| TreeWith60ListItems.perf.tsx | 135.24 | 143.21 | 141.79 | 175.94 |
| TreeWith60ListItems.perf.tsx | 136.88 | 142.5  | 140.67 | 177    |
### `same-page`
| Example                      | min   | avg    | median | max    |
| ---------------------------- | ----- | ------ | ------ | ------ |
| ChatWithPopover.perf.tsx     | 93.71 | 105.98 | 98.85  | 200.73 |
| ChatWithPopover.perf.tsx     | 98.78 | 115.03 | 106.32 | 217.35 |
| DropdownManyItems.perf.tsx   | 96.13 | 111.63 | 105.56 | 206.32 |
| DropdownManyItems.perf.tsx   | 95.76 | 107.7  | 103.16 | 198.6  |
| TreeWith60ListItems.perf.tsx | 47.56 | 57.13  | 54.51  | 115.5  |
| TreeWith60ListItems.perf.tsx | 47.23 | 58.66  | 53.23  | 133.9  |

</details>