@layershifter 2020-06-02
> Could you add a perf example measurements to see the perf impact for consumers?

Added 👍